+++ CHANGELOG.md
@@ -1,3 +1,7 @@
+# Head
+
+- Fixed: `time-no-imperceptible` correctly handles negative time.
+
 # 6.8.0
 
 - Deprecated: `-e` and `--extract` CLI flags, and the `extractStyleTagsFromHtml` node API option. If you use these flags or option, please consider creating a processor for the community. See the [release planning](/docs/user-guide/release-planning.md) document for more details.